+++ pyprocessmacro/__init__.py
@@ -3,7 +3,8 @@
 Core module for the PyProcess Macro.
 """
 from .process import Process
+from .utils import ConvergenceError
 
-__all__ = ["Process"]
+__all__ = ["Process", "ConvergenceError"]
 
 __version__ = "1.0.14"